+++ Fwk/VizFwk/LibGeometry/cvfBoundingBoxTree.cpp
@@ -485,7 +485,7 @@ bool AABBTree::buildTree()
     {
         bool bThreadRes = bRes;
 #pragma omp for
-        for (int i = 0; i < m_previousLevelNodes.size(); ++i)
+        for (int i = 0; i < static_cast<int>(m_previousLevelNodes.size()); ++i)
         {
             bThreadRes = bThreadRes && buildTree(m_previousLevelNodes[i].node, m_previousLevelNodes[i].fromIdx, m_previousLevelNodes[i].toIdx, 3, 6);
         }
@@ -500,7 +500,7 @@ bool AABBTree::buildTree()
     {
         bool bThreadRes = bRes;
 #pragma omp for schedule(guided)
-        for (int i = 0; i < m_previousLevelNodes.size(); ++i)
+        for (int i = 0; i < static_cast<int>(m_previousLevelNodes.size()); ++i)
         {
             bThreadRes = bThreadRes && buildTree(m_previousLevelNodes[i].node, m_previousLevelNodes[i].fromIdx, m_previousLevelNodes[i].toIdx, 6);
         }
